backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Subhash Chandra Shukla <subhashdev011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
pydantic = "^2.11.2"
uvicorn = "^0.34.0"
motor = "^3.7.0"
bcrypt = "^4.3.0"
cryptography = "^44.0.2"
python-jose = "^3.4.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
passlib = "^1.7.4"
pydantic-settings = "^2.8.1"
dotenv = "^0.9.9"
email-validator = "^2.2.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
